Add relace old_domain with new_domain to unload-membership
